requirements.txt
anthropic>=0.18.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
PyPDF2>=3.0.1
tiktoken>=0.5.2
numpy>=1.24.0
tqdm>=4.66.0
